% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_persist.R
\name{build_persist}
\alias{build_persist}
\title{Build rasters of persistent refugia}
\usage{
build_persist(
  esm = c("ens", "gfdltv", "hadtv", "ipsltv"),
  yr_range = c(1990:2019),
  persist_thresh = c(50, 95),
  save_path = NULL
)
}
\arguments{
\item{esm}{Character. ESM model to use. Choose from \code{ens}, \code{gfdltv}, \code{hadtv}, \code{ipsltv}.}

\item{yr_range}{Vector of integers. Time period.}

\item{persist_thresh}{Integer. Temporal threshold (\%) used for defining annual refugia. Choose either 50 (liberal) or 95\% (conservative).}

\item{save_path}{Character. Directory to save output to. Optional, set to NULL if not needed.}
}
\value{
Produces a \code{terra::rast()} object of persistent refugia (\%), for the ESM specified.
}
\description{
This function creates a single raster of persistent refugia for an ESM of interest,
over a user-specified time period.
}
\examples{
persist_refugia <- build_persist(esm = "ens", yr_range = 2070:2099,
persist_thresh = 50, save_path = NULL)
persist_refugia
terra::plot(persist_refugia)
}
