% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_refugia_rasters.R
\name{build_refugia_rasters}
\alias{build_refugia_rasters}
\title{Build rasters of binary refugia}
\usage{
build_refugia_rasters(
  percentdays = abalone::percentdays,
  esm = c("ens", "gfdltv", "hadtv", "ipsltv"),
  yrst = 1990,
  yrend = 2100,
  thresh = c(50, 95),
  progress = TRUE
)
}
\arguments{
\item{percentdays}{R object. Dataframe. Defaults to \code{abalone::percentdays} (i.e., example package dataset of \% days of stress)}

\item{esm}{Character. ESM model to use. Choose from \code{ens}, \code{gfdltv}, \code{hadtv}, \code{ipsltv}.}

\item{yrst}{Integer. Start year. Defaults to 1990.}

\item{yrend}{Integer. End year. Defaults to 2100.}

\item{thresh}{Integer. Temporal threshold (\%) used for defining annual refugia. Choose either 50 (liberal) or 95\% (conservative).}

\item{progress}{Logical. Show progress bar? Default is TRUE.}
}
\value{
Produces a \code{terra::rast()} object of binary refugia (0/1), for the ESM specified, for all years between \code{yrst} and \code{yrend}.
}
\description{
This function creates annual rasters of refugia for an ESM of interest, using a
temporal threshold (where refugia conditions must be met for 50 or 95\% of the year).
The function also displays a progress bar.
}
\examples{
ens_refugia <- build_refugia_rasters(percentdays = abalone::percentdays,
esm = "ens", yrst = 1990, yrend = 2100, thresh = 50, progress = FALSE)
ens_refugia
terra::plot(ens_refugia[[98]], main = "Refugia in 2088")


}
