% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali_rast.R
\name{cali_rast}
\alias{cali_rast}
\title{ROMS 3km California subsetted to 100m depth}
\source{
Custom defined by project team
}
\usage{
cali_rast(
  pth = system.file("extdata", "emptyrast_100.tif", package = "abalone")
)
}
\arguments{
\item{pth}{Internal package file path to where the \code{.tif} is stored}
}
\description{
This is a \code{terra::rast} object of the California coast, subsetted to only include cells above 100 m depth.
}
\examples{
emptyrast_100 <- cali_rast()
emptyrast_100
terra::plot(emptyrast_100)

}
