% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltarast.R
\name{deltarast}
\alias{deltarast}
\title{Build rasters of persistent refugia}
\usage{
deltarast(
  thresh = c(50, 95),
  model = c("ens", "gfdltv", "hadtv", "ipsltv"),
  def = "def8",
  histrange = 2020:2049,
  projrange = 2070:2099,
  save_path = NULL
)
}
\arguments{
\item{thresh}{Integer. Temporal threshold (\%) used for defining annual refugia. Choose either 50 (liberal) or 95\% (conservative).}

\item{model}{Character. ESM model to use. Choose from \code{ens}, \code{gfdltv}, \code{hadtv}, \code{ipsltv}.}

\item{def}{Character. Refugia definition name (e.g., "def8") used in input file paths.}

\item{histrange}{Vector of integers. Time period for historical period}

\item{projrange}{Vector of integers. Time period for projection period}

\item{save_path}{Character. Directory to save output to. Optional, set to NULL if not needed.}
}
\value{
Produces a \code{terra::rast()} object of delta refugia (\%), for the ESM specified.
}
\description{
This function creates a single raster of persistent refugia for an ESM of interest,
over a user-specified time period.
}
\examples{
deltar <- deltarast(thresh = 50, model = "ens", histrange = 2020:2049,
projrange = 2070:2099, save_path = NULL)
deltar
terra::plot(deltar)
}
