% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extent_list.R
\docType{data}
\name{extent_list}
\alias{extent_list}
\title{List of extents for case study regions}
\format{
An object of class \code{list} of length 4.
}
\source{
Custom defined by project team
}
\usage{
extent_list
}
\description{
This object contains bounding boxes (latitudes and longitudes) for key study regions used in the refugia analysis. These can be used with \code{terra::ext()} for vizualisation.
}
\details{
\strong{Available extents:}
\itemize{
\item \strong{Monterey Bay}: \code{monterey_bay}
\item \strong{Fort Bragg}: \code{fort_bragg}
\item \strong{Channel Islands}: \code{channel_islands}
\item \strong{San Francisco}: \code{san_francisco}
}
}
\examples{
extent_list
}
\keyword{datasets}
