% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentdays_ts.R
\name{percentdays_ts}
\alias{percentdays_ts}
\title{Plot percent of days meeting refugia conditions over time}
\usage{
percentdays_ts(
  area = c("monterey_bay", "channel_islands", "fort_bragg", "san_francisco"),
  yr_range = 1990:2100,
  def = c("def3", "def4", "def5", "def6", "def7", "def8"),
  input_dir = "/Users/admin/Documents/GitHub/savingabalone/out",
  cons_thresh = 95,
  lib_thresh = 50
)
}
\arguments{
\item{area}{Character. Name of the area; must match a key in \code{abalone::extent_list}.}

\item{yr_range}{Numeric vector. Range of years to include. Default is 1990–2100.}

\item{def}{Character. Refugia definition name (e.g., "def5") used in input file paths.}

\item{input_dir}{Character. Parent directory where input files live. Defaults to \verb{/out}}

\item{cons_thresh}{Integer. Value representing the conservative temporal threshold to define refugia. Defaults to 95(\%)}

\item{lib_thresh}{Integer. Value representing the liberal temporal threshold to define refugia. Defaults to 50(\%)}
}
\value{
Produces a \code{ggplot2::ggplot} object of an annual time series of refugia.
}
\description{
This function produces a \code{ggplot2} object of an annual time series of refugia
across all ESMs, including the ensemble mean. The plot includes dashed horizontal
lines for the two temporal thresholds used to define refugia: 95 and 50\%. Note
this function will only work for refugia definitions that are calculated using
\% of year defined as refugia (i.e., not Definitions 1-2).
}
\examples{
percentdays_ts(area = "monterey_bay", yr_range = 1990:2100, def = "def5", input_dir = "/Users/admin/Documents/GitHub/savingabalone/out", cons_thresh = 95, lib_thresh = 50)

}
