% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shp.R
\name{read_shp}
\alias{read_shp}
\title{Read in shapefiles provided by this package}
\usage{
read_shp(shape = c("usa", "eez", "contshelf50", "contshelf200", "mpas"))
}
\arguments{
\item{shape}{One of the default shapefiles to use. Options are:
\itemize{
\item \code{"usa"} – The contiguous United States land boundary.
\item \code{"eez"} – The U.S. West Coast EEZ
\item \code{"contshelf50"} – The 50 m isobath
\item \code{"contshelf200"} – The 200m isobath (i.e., the continental shelf)
\item \code{"mpas"} – Californian marine protected areas
}}
}
\value{
An \code{sf} object of the chosen shapefile
}
\description{
This function uses the \code{sf} package to read one of the included shapefiles bundled with the package. Jessie has standardized these.
}
\details{
The shapefiles are stored in \code{/extdata}.

\strong{Available shapefiles:}
\itemize{
\item \strong{usa}: \code{usa_contiguous.shp}
\item \strong{eez}: \code{westcoast_eez.shp}. Derived from 'Marine Regions' dataset: https://www.marineregions.org/
\item \strong{contshelf50}: \code{cont_shelf_50m.shp}. Derived from GEBCO 2024
\item \strong{contshelf200}: \code{cont_shelf_200m.shp}. Derived from GEBCO 2024.
\item \strong{mpas}: \code{california_mpas.shp}. From CDFW dataset: 'California Marine Protected Areas \link{ds582}', representing all MPAs as of 1 Jan, 2019.
}
}
