% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stress_rasters.R
\name{build_stress_rasters}
\alias{build_stress_rasters}
\title{Create annual stress \% rasters}
\usage{
build_stress_rasters(
  percentdays = abalone::percentdays,
  esm = c("ens", "gfdltv", "hadtv", "ipsltv"),
  yrst = 1990,
  yrend = 2100,
  progress = TRUE,
  save_path = NULL
)
}
\arguments{
\item{percentdays}{R object. Dataframe. Defaults to \code{abalone::percentdays} (i.e., example package dataset of \% days of stress)}

\item{esm}{Character. ESM model to use. Choose from \code{ens}, \code{gfdltv}, \code{hadtv}, \code{ipsltv}.}

\item{yrst}{Integer. Start year.}

\item{yrend}{Integer. End year.}

\item{progress}{Logical. Show progress bar? Default is TRUE.}

\item{save_path}{Character. Directory to save output to. Optional, set to NULL if not needed.}
}
\value{
Produces a \code{terra::rast()} object of \% of year stressed, for the ESM specified, for all years between \code{yrst} and \code{yrend}.
}
\description{
This function produces a \code{terra:rast()} object containing annual stress rasters, defined as
the percentage of each year that abalone experience stress based on the definition chosen.
}
\examples{
ens_stress <- build_stress_rasters(percentdays = abalone::percentdays, esm = "ens",
yrst = 1990, yrend = 1992, progress = FALSE, save_path = NULL)
ens_stress
terra::plot(ens_stress[[2]], main = "\% stress in 1992")

}
